requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
dnspython==2.7.0
fastapi==0.116.1
graphql-core==3.2.6
h11==0.16.0
idna==3.10
lia-web==0.2.3
motor==3.7.1
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
strawberry-graphql==0.280.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
